added character selection

+++ src/main/java/olter/loaf/game/games/dto/PublicPlayerResponse.java
@@ -14,5 +14,5 @@ public class PublicPlayerResponse {
     private Integer gold;
     private Integer handSize;
     private List<PlayerDistrictResponse> districts;
-    private Long currentCharacter;
+    private Integer currentCharacter;
 }
